% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/followup.R
\name{followup_priorities}
\alias{followup_priorities}
\title{Compute the followup priorities for a list of contacts}
\usage{
followup_priorities(contact_list, exposure, exposure_end = NULL,
  rate_infectious_contact = NULL, last_followup = NULL,
  p_disease = 1, incubation_period = NULL,
  date_analysis = Sys.Date(), sort = TRUE)
}
\arguments{
\item{contact_list}{A data.frame with one row per contact, containing at
least a column with possible dates of exposure.}

\item{exposure}{The name of the column of contact_list containing the dates
of exposure (variable name or in quotes). Can be a list column
containing vectors with several possible exposure dates per contact.}

\item{exposure_end}{The name of a column containing dates representing the
end of the exposure period. This is \code{NULL} by default, indicating all
exposures are in the \code{exposure} column.}

\item{rate_infectious_contact}{a vector containing rates of infectious contact corresponding
to each exposure date in exposure, or between exposure and exposure_end.
same length as exposure. Defaults to rep(1, length(exposure)).}

\item{last_followup}{The name of the column of contact_list containing the
last follow up date for each contact (variable name or in quotes).
Should contain NA if unknown or contact has never been followed.}

\item{incubation_period}{The incubation period distribution. Can be a
distcrete distribution, an empirical incubation period returned by
epitrix::empirical_incubation_dist() (of type data.frame) or an atomic
vector whose elements 1:n correspond to the probability of the incubation
period being 0:(n-1).}

\item{date_analysis}{the date on which the prioritization should be done. The
probability of symptoms onset is computed until the end of the previous
day. Defaults to the system date (i.e. today).}

\item{sort}{If TRUE (default) the result is sorted by followup priority.}
}
\value{
A data.frame containing all columns in contact_list and the following
additional columns:
\itemize{
\item rate_onset the overall rate of disease onset by
date_analysis
\item p_onset the overall probability of disease onset by
date_analysis
\item followup_priority, an index ranging from 1 for the highest
priority to the number of patients
}
}
\description{
Outputs the probability that the symptoms onset of a contact was between the
last follow up (day of last follow up not included in interval)
and the analysis date, which defaults to
day before current system date and is included in interval.
}
