% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getContrasts.R
\encoding{UTF-8}
\name{getContrasts}
\alias{getContrasts}
\alias{getTrends}
\title{Linear Contrasts and Polynomial Trends for ANOVAs}
\usage{
getContrasts(mod, factor, contr, split = NULL, pooled = FALSE,
             adj = 'none')

getTrends(mod, factor, split = NULL, pooled = FALSE, adj = 'none')
}
\arguments{
\item{mod}{An \code{afex_aov} ANOVA model to use for contrast analysis.}

\item{factor}{\code{character} vector (of length 1) indicating the name of
the factor to perform contrasts on in the provided ANOVA model.}

\item{contr}{\code{list} of contrasts to perform on the model.}

\item{split}{\code{character} vector indicating the name of the factor (if
any) to split the analysis on. Defaults to \code{NULL} (no split).}

\item{pooled}{\code{logical} indicating whether to use the pooled
error term and degrees of freedom from the unsplit model for increased
statistical power when splitting a contrast. Defaults to \code{FALSE}.}

\item{adj}{\code{character} vector indicating which method of p-value
adjustment for multiple comparisons should be used when generating the 
output table (e.g. "tukey", "bonferroni"). Defaults to "none".}
}
\value{
A \code{\link[emmeans]{summary_emm}} and \code{\link{data.frame}}
  containing the contrast or trend results.
}
\description{
These functions perform linear contrast and polynomial trend analyses on 
ANOVAs and ANCOVA models produced by the \pkg{afex} package, providing 
mean-squared errors, F-statistics, and (partial/contrast) eta-squared 
effect sizes in a manner similar to SPSS. They serve as a wrapper around
the \code{emmeans::contrast} method from the \pkg{emmeans} package, which
only reports t.scores for contrasts and does not provide effect sizes or MSE.
Splitting the contrasts across levels of a factor is supported, as is using
the pooled error term from the omnibus model when splitting on a
between-subjects factor.
}
\author{
Austin Hurst
}
