% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_new_onset.R
\name{p_new_onset}
\alias{p_new_onset}
\title{Computes the probability of new symptoms on a particular day for a particular contact}
\usage{
p_new_onset(incubation_period, date_analysis, exposure)
}
\arguments{
\item{incubation_period}{a vector of probabilities. Has to sum to 1.}

\item{date_analysis}{the date for which to compute the probability of new
symptoms}

\item{exposure}{a vector containing one or several possible dates of
exposure. They are equally weighted.}
}
\value{
a probability
}
\description{
Computes the probability of new symptoms on a particular day for a particular contact
}
