% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleEffects.R
\encoding{UTF-8}
\name{simpleEffects}
\alias{simpleEffects}
\title{Simple Effects Analysis for Factorial ANOVAs}
\usage{
simpleEffects(mod, split, pooled = FALSE, returns = 'summary',
  correction = 'GG')
}
\arguments{
\item{mod}{An \code{afex_aov} ANOVA model returned by \code{aov_car}, \code{aov_4},
or \code{aov_ez}.}

\item{split}{A \code{character} vector (of length 1) indicating the name
of the factor the in the ANOVA model to split the analysis on.}

\item{returns}{\code{character} vector indicating the format in which to 
return the results. Can be 'summary', 'table', 'nice', or 'afex_aov'.
'table' and 'nice' are concise and easy to read, but omit sphericity
test information for repeated-measures designs. 'afex_aov' returns the
full afex anova object. Defaults to 'summary'.}

\item{pooled}{\code{logical} indicating whether to use the pooled error term
and degrees of freedom from the unsplit model for increased statistical
power when splitting a model. Defaults to \code{FALSE}.}

\item{correction}{\code{character} vector specifying which method of
sphericity correction to use for repeated-measures factors, can be 'GG'
(Greenhouse-Geisser), 'HF' (Huynh-Feldt), or 'none'. Defaults to 'GG'.}
}
\value{
A \code{\link{list}} of ANOVA tables. The format of the tables
  depends on the value of the \code{returns} parameter.
}
\description{
This function performs simple effects analysis on ANOVA models, allowing
easy splitting of models by a given factor. Additionaly, this function
supports using the pooled error term and degrees of freedom from the 
unsplit model for increased statistical power.
}
\author{
Austin Hurst
}
