% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_html.R
\name{to_html}
\alias{to_html}
\title{Helper function print contact list as html table.}
\usage{
to_html(contact_list, cols_round = list(), date_cols = list(),
  ndigit = 2)
}
\arguments{
\item{contact_list}{a data frame}

\item{cols_round}{a list of quoted column names to be rounded}

\item{date_cols}{a list of quoted column names containing dates}

\item{ndigit}{number of digits for rouning (default: 2)}
}
\description{
Helper function print contact list as html table.
}
