% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/followup.R
\name{followup_priorities}
\alias{followup_priorities}
\title{Compute the followup priorities for a list of contacts}
\usage{
followup_priorities(contact_list, exposure, exposure_end = NULL,
  last_followup = NULL, p_disease = 1, incubation_period = NULL,
  date_analysis = Sys.Date(), include_last_follow_up = TRUE,
  sort = TRUE)
}
\arguments{
\item{contact_list}{A data.frame with one row per contact, containing at
least a list column with possible dates of exposure.}

\item{exposure}{The name of the column of contact_list containing the dates
of exposure (bare variable name or in quotes). Can be a list column
containing vectors with several possible exposure dates per contact.}

\item{exposure_end}{the name of a column containing dates representing the
end of the exposure period. This is \code{NULL} by default, indicating all
exposures are known and in the \code{exposure} column.}

\item{last_followup}{The name of the column of contact_list containing the
last follow up date for each contact (bare variable name or in quotes).
Should contain NA if unknown or contact has never been followed.}

\item{p_disease}{The overall probability that exposure leads to disease. Can
either be a scalar applying to all contacts, or the name of a column (bare
variable name or in quotes) in contact_list containing a different
probability per contact. Defaults to 1.}

\item{incubation_period}{The incubation period distribution. Can be a
distcrete distribution, an empirical incubation period returned by
epitrix::empirical_incubation_dist() (of type data.frame) or an atomic
vector whose elements 1:n correspond to the probability of the incubation
period being 0:(n-1).}

\item{date_analysis}{the date on which the prioritization should be done. The
probability of symptoms onset is computed until the end of the previous
day. Defaults to the system date (i.e. today).}

\item{include_last_follow_up}{TRUE if the date of last follow up should be
included in the date range and thus the probability of symptoms starting on
that date included in the result, FALSE if not (default TRUE).}

\item{sort}{If TRUE (default) the result is sorted by followup priority.}
}
\value{
A data.frame containing all columns in contact_list and the following
additional columns:
\itemize{
\item p_onset, the probability of disease has onset by date_analysis
given that the exposure results in disease
\item p_symptoms, the overall probability of disease has onset by
date_analysis
\item followup_priority, an index ranging from 1 for the highest
priority to the number of patients
}
}
\description{
Outputs the probability that the symptoms onset of a contact was between the
last follow up (day of last follow up included in interval by default, see
parameter include_last_follow_up) and the analysis date, which defaults to
day before current system date and is included in interval.
}
