% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_integral_onset.R
\name{p_integral_onset}
\alias{p_integral_onset}
\title{Computes the probability of new symptoms within a range of dates for a
particular contact, given the incubation period.}
\usage{
p_integral_onset(incubation_period, date_lower, date_upper, exposure,
  exposure_weight = NULL)
}
\arguments{
\item{incubation_period}{a vector of probabilities. Has to sum to 1.}

\item{date_lower}{the lower limit for which to compute the probability of new
symptoms.}

\item{date_upper}{the upper limit for which to compute the probability of new
symptoms.}

\item{exposure}{a vector containing one or several possible dates of
exposure. They are equally weighted.}

\item{exposure_weight}{a vector containing weights applied to different exposure dates.
same length as exposure. Defaults to rep(1, length(exposure)).}
}
\value{
a probability
}
\description{
The lower but not the upper limit is included in the interval such that the
probability returned is the sum over the probability for each date_analysis
that fulfils the following: date_lower <= date_analysis < date_upper.
}
